fix: no need to move to double for amount

+++ sample/src/main/java/com/airwallex/paymentacceptance/repo/PACheckoutDemoRepository.kt
@@ -41,7 +41,7 @@ class PACheckoutDemoRepository : BaseRepository {
             "apiKey" to Settings.apiKey,
             "clientId" to Settings.clientId,
             "request_id" to UUID.randomUUID().toString(),
-            "amount" to (amount ?: Settings.price.toBigDecimal()).toDouble(),
+            "amount" to (amount ?: Settings.price.toBigDecimal()),
             "currency" to Settings.currency,
             "merchant_order_id" to UUID.randomUUID().toString(),
             "order" to PurchaseOrder.Builder()
